task 6

+++ 0x00-hello_world/6-size.c
@@ -13,15 +13,15 @@ float c;
 long int d;
 long long int e;
 
-printf("Size of char is: %lu byte(s)", sizeof(a));
+printf("Size of char is: %lu byte(s)",(unsigned long)sizeof(a));
 
-printf("Size of int is: %lu byte(s)", sizeof(b));
+printf("Size of int is: %lu byte(s)", (unsigned long)sizeof(b));
 
-printf("Sizeof long int is : %lu byte(s)", sizeof(d));
+printf("Sizeof long int is : %lu byte(s)", (unsigned long)sizeof(d));
 
-printf("Size of float is : %d byte(s)" sizeof(c));
+printf("Size of float is : %lu byte(s)" (unsigned long)sizeof(c));
 
-printf("Size of long long int is : %d byte(s)", sizeof(e));
+printf("Size of long long int is : %lu byte(s)", (unsigned long)sizeof(e));
 return (0);
 
 }
